# docs/swagger.yaml
definitions:
...
  models.CreateGatewayscReq:
    properties:
      active:
        type: integer
      createdDate:
        type: string
      gateway_description:
        type: string
      gateway_image:
        type: string
      gateway_type:
        type: string
      gateway_url:
        type: string
      id:
        type: string
      updatedDate:
        type: string
    type: object
  models.CreateIngresoReq:
    properties:
      activo:
        type: integer
      author:
        type: string
      company_id:
        type: integer
      createdDate:
        type: string
      curso_id:
        type: integer
      fecha:
        type: string
      fechainicial:
        type: string
      fpago:
        type: string
      id:
        type: string
      identificador:
        type: string
      monto:
        type: number
      nrocuotas:
        type: integer
      reserva:
        type: integer
      rutalum:
        type: string
      rutapo:
        type: string
      sale_id:
        type: integer
      status_pago:
        type: string
      tipocomp:
        type: string
      token_flow:
        type: string
      updatedDate:
        type: string
      valorcuota:
        type: number
    type: object
...
  models.GatewayscResp:
    properties:
      active:
        type: integer
      createdDate:
        type: string
      gateway_description:
        type: string
      gateway_image:
        type: string
      gateway_type:
        type: string
      gateway_url:
        type: string
      id:
        type: string
      updatedDate:
        type: string
    type: object
...
  models.IngresoResp:
    properties:
      activo:
        type: integer
      author:
        type: string
      company_id:
        type: integer
      createdDate:
        type: string
      curso_id:
        type: integer
      fecha:
        type: string
      fechainicial:
        type: string
      fpago:
        type: string
      id:
        type: string
      identificador:
        type: string
      monto:
        type: number
      nrocuotas:
        type: integer
      reserva:
        type: integer
      rutalum:
        type: string
      rutapo:
        type: string
      sale_id:
        type: integer
      status_pago:
        type: string
      tipocomp:
        type: string
      token_flow:
        type: string
      updatedDate:
        type: string
      valorcuota:
        type: number
    type: object
...
  models.SaleResp:
    properties:
      accesscode:
        type: string
      activo:
        type: integer
      author:
        type: string
      comision:
        type: number
      comision_pagada:
        type: integer
      company_id:
        type: integer
      correo_encargado:
        type: string
      createdDate:
        type: string
      cuotas:
        type: integer
      curso:
        type: integer
      descm:
        type: integer
      descrition:
        type: string
      encargado:
        type: string
      establecimiento_id:
        type: integer
      fecha:
        type: string
      fecha_cierre:
        type: string
      fecha_ultpag:
        type: string
      fechacuota:
        type: string
      fechasalida:
        type: string
      id:
        type: string
      idcurso:
        type: string
      identificador:
        type: string
      liberados:
        type: integer
      nroalumno:
        type: integer
      obs:
        type: string
      password:
        type: string
      program_id:
        type: integer
      seller_id:
        type: integer
      sendemail:
        type: integer
      state:
        type: string
      subtotal:
        type: integer
      tipocambio:
        type: number
      type_sale:
        type: string
      updatedDate:
        type: string
      vprograma:
        type: integer
    type: object
...
  models.UpdateUsersReq:
    properties:
      active:
        type: integer
      author:
        type: string
      company_id:
        type: integer
      createdDate:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      reset_token:
        type: string
      reset_token_expira:
        type: string
      roles_id:
        type: integer
      updatedDate:
        type: string
      username:
        type: string
    type: object
...
  models.UsersResp:
    properties:
      active:
        type: integer
      author:
        type: string
      company_id:
        type: integer
      createdDate:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      reset_token:
        type: string
      reset_token_expira:
        type: string
      roles_id:
        type: integer
      updatedDate:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: Powered by scv-go-tools - https://github.com/sergicanet9/scv-go-tools
  title: Go Hexagonal API
paths:
...
  /api/v3.5/Gatewaysc:
    get:
      description: Gets all the Gatewaysc
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GatewayscResp'
            type: array
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get all Gatewaysc
      tags:
      - Gatewaysc
    post:
      description: Creates a new Gatewaysc
      parameters:
      - description: New colegio to be created
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateGatewayscReq'
      responses:
        "201":
          description: OK
          schema:
            $ref: '#/definitions/models.GatewayscResp'
        "400":
          description: Bad Request
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Create Gatewaysc
      tags:
      - colegion
  /api/v3.5/Gatewaysc/{id}:
    delete:
      description: Delete a Gatewaysc
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Delete Gatewaysc
      tags:
      - Gatewaysc
    get:
      description: Gets a Gatewaysc by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GatewayscResp'
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get Gatewaysc by ID
      tags:
      - Gatewaysc
...
  /api/v3.5/ingreso:
    get:
      description: Gets all the ingreso
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SaleResp'
            type: array
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get inf all ingreso
      tags:
      - ingreso
    post:
      description: Creates a new ingreso
      parameters:
      - description: New colegio to be created
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateIngresoReq'
      responses:
        "201":
          description: OK
          schema:
            $ref: '#/definitions/models.IngresoResp'
        "400":
          description: Bad Request
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Create ingreso
      tags:
      - ingreso
  /api/v3.5/ingreso/{id}:
    delete:
      description: Delete a ingreso
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Delete ingreso
      tags:
      - ingreso
    get:
      description: Gets a ingreso by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IngresoResp'
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get ingreso by ID
      tags:
      - ingreso
...
  /api/v3.5/users/{id}:
    delete:
      description: Delete a users
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Delete users
      tags:
      - users
    get:
      description: Gets a users by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UsersResp'
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Get users by ID
      tags:
      - users
    patch:
      description: Updates a users
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: users
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUsersReq'
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "408":
          description: Request Timeout
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - Bearer: []
      summary: Update users
      tags:
      - users
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"